+++ src/test/java/io/github/caffetteria/data/service/opencmis/TestOpencmisConfig.java
@@ -1,5 +1,7 @@
 package io.github.caffetteria.data.service.opencmis;
 
+import org.fugerit.java.simple.config.ConfigParams;
+import org.fugerit.java.simple.config.ConfigParamsDefault;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -14,7 +16,7 @@ class TestOpencmisConfig {
         Properties configProperties = new Properties();
         String userTest = "user1";
         configProperties.setProperty( OpencmisDataServiceFacade.KEY_USER, userTest );
-        OpencmisDataServiceConfig config = new OpencmisDataServiceConfigDefault( configProperties );
+        ConfigParams config = new ConfigParamsDefault( configProperties );
         Optional<String> testFound = config.getOptionalValue( OpencmisDataServiceFacade.KEY_USER );
         Optional<String> testNotFound = config.getOptionalValue( "notFound" );
         Assertions.assertEquals( userTest, testFound.get() );